releng: add simple Nix expression for devshell
Nix[1] is an alternative package manager that you can
install in parallel with your system package manager
on Linux or MacOS. It allows you to enter a development
environment with all necessary dependencies installed
simply by running:

    nix-shell

We use package definitions from Nix package repository
and ROS overlay, which we are pinning using experimental
Nix flake mechanism. Using unstable Nix is not necessary,
thanks to a compatibility shim, except for updating
the lock file itself. But you can easily run that command
with stable Nix as well:

    nix-shell -I nixpkgs=channel:nixos-unstable -p nixUnstable --run 'nix --experimental-features "nix-command flakes" flake update'

And if keeping up with flakes turns out to be too much hassle,
it should be quite easy to switch to niv[4] for pinning.

Also note that currently, all dependencies for `desktop` and `doc` suites
are included. In the future, we might want to load them more lazily.

Lastly, beware that running graphical applications that depend on OpenGL
on non-NixOS Linux distros will require starting them using nixGL[5].

[1]: https://nixos.org/
[2]: https://github.com/NixOS/nixpkgs
[3]: https://github.com/lopsided98/nix-ros-overlay
[4]: https://github.com/nmattia/niv
[5]: https://github.com/guibou/nixGL