rofi-app: fix build with VTK 9
Also fixes the runtime warnings but not the CMake warnings since those are harder to work around.

I used QVTKOpenGLNativeWidget since unlike QVTKOpenGLStereoWidget (called QVTKOpenGLWidget in 8.2), it is called the same in 9.0 and it is also recommended. Though, to add to the confusion the native widget was actually called QVTKOpenGLWidget before 8.2 so maybe it was intended to be used in 08b7aee4c9acb28e6e2068b0326bd4a3034cd247.